# Conflicts:
#	.gitlab-ci.yml
#	gitential2/backends/in_memory.py
#	poetry.lock
#	pyproject.toml

+++ gitential2/cli_v2/workspaces.py
@@ -5,9 +5,11 @@ from structlog import get_logger
 
 from gitential2.core.users import get_user
 from .common import get_context
+from ..backends.sql.cleanup import perform_data_cleanup
 from ..core.api_keys import delete_api_keys_for_workspace
 from ..core.workspace_common import duplicate_workspace
 from ..datatypes import UserInDB, WorkspaceMemberInDB
+from ..datatypes.cli_v2 import ResetType, CleanupType
 from ..datatypes.workspaces import WorkspaceDuplicate
 from ..exceptions import SettingsException
 
@@ -16,7 +18,7 @@ logger = get_logger(__name__)
 
 
 @app.command("reset")
-def reset_workspace(workspace_id: int):
+def reset_workspace(workspace_id: int, reset_type: ResetType = typer.Option("full", "--type", "-t")):
     """
     DANGER ZONE!!! Workspace reset!
     \n
@@ -31,8 +33,10 @@ def reset_workspace(workspace_id: int):
     workspace = g.backend.workspaces.get(id_=workspace_id) if workspace_id else None
     if workspace:
         logger.info("Starting to truncate all of the tables for workspace!", workspace_id=workspace.id)
-        g.backend.reset_workspace(workspace_id=workspace_id)
-        g.kvstore.delete_values_for_workspace(workspace_id=workspace_id)
+        if reset_type in (ResetType.full, ResetType.sql_only):
+            g.backend.reset_workspace(workspace_id=workspace_id)
+        if reset_type in (ResetType.full, ResetType.redis_only):
+            g.kvstore.delete_values_for_workspace(workspace_id=workspace_id)
     else:
         logger.exception("Failed to reset workspace! Workspace not found by the provided workspace id!")
 
@@ -107,3 +111,29 @@ def purge_workspace(workspace_id: int):
             logger.exception("Failed to purge workspace! Can not purge primary workspace!")
     else:
         logger.exception("Failed to purge workspace! Workspace not found by the provided workspace id!")
+
+
+@app.command("cleanup")
+def perform_workspace_cleanup(
+    workspace_id: int = typer.Argument(None), cleanup_type: CleanupType = typer.Option("full", "--type", "-t")
+):
+    """
+    \b
+    This command will delete all redundant data from the PostgreSQL database and also from Redis.
+    Data is considered to be redundant if it is not used for anything. Like when a project is deleted and
+    the commits for a repo just lies there without any reason. The same thing happens with the Redis too.
+    """
+
+    g = get_context()
+    workspace = g.backend.workspaces.get(id_=workspace_id) if workspace_id else None
+
+    if workspace_id and not workspace:
+        logger.exception(
+            "Failed to cleanup workspace! Workspace not exists for given workspace id!", workspace_id=workspace_id
+        )
+        return
+
+    if workspace:
+        perform_data_cleanup(g=g, workspace_ids=[workspace.id], cleanup_type=cleanup_type)
+    else:
+        perform_data_cleanup(g=g, workspace_ids=[w.id for w in g.backend.workspaces.all()], cleanup_type=cleanup_type)
